+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.2.1
+
+* Normal and visual keys can be defined in .keymap instead of .behaviors
+  * Use :editor.keys.vim.normal.cm tag in keymap instead of deprecated lt.plugins.vim/map-keys behavior
+    For example, `[:app :lt.plugins.vim/map-keys {"j" "gj" "k" "gk"}]` in a behaviors file,
+    is `[:editor.keys.vim.normal.cm "j" "gj"] [:editor.keys.vim.normal.cm "k" "gk"]` in a keymap file.
+  * Use :editor.keys.vim.visual.cm tag in keymap instead of deprecated lt.plugins.vim/map-keys-visual behavior
+* Add :qa ex command
+
 ## 0.2.0
 
 * Plugin activated by default. :lt.plugins.vim/activate-vim behavior in